% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hima_data.R
\docType{data}
\name{BinaryOutcome}
\alias{BinaryOutcome}
\title{Binary Outcome Dataset for HIMA Demo}
\format{
A list with the following components:
\describe{
  \item{PhenoData}{A data frame containing:
    \describe{
      \item{Treatment}{treated (value = 1) or not treated (value = 0).}
      \item{Disease}{binary outcome: diseased (value = 1) or healthy (value = 0).}
      \item{Sex}{female (value = 1) or male (value = 0).}
      \item{Age}{age of the participant.}
    }
  }
  \item{Mediator}{A matrix of high-dimensional mediators (rows: samples, columns: variables).}
}
}
\usage{
BinaryOutcome
}
\description{
A dataset containing phenotype data and high-dimensional mediators for binary outcome analysis.
The dataset was simulated using parameters generated from real data.
}
\examples{
data(BinaryOutcome)
head(BinaryOutcome$PhenoData)

}
\keyword{datasets}
