% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dblassoHIMA.R
\name{dblassoHIMA}
\alias{dblassoHIMA}
\title{This is the function for high-dimensional mediation analysis using de-biased lasso
HIMA with de-biased lasso}
\usage{
dblassoHIMA(
  X,
  Y,
  M,
  Z,
  Y.family = c("gaussian", "binomial"),
  topN = NULL,
  scale = TRUE,
  FDRcut = 0.05,
  verbose = FALSE
)
}
\arguments{
\item{X}{a vector of exposure.}

\item{Y}{a vector of outcome. Can be either continuous or binary (0-1).}

\item{M}{a \code{data.frame} or \code{matrix} of high-dimensional mediators. Rows represent samples, columns represent variables.}

\item{Z}{a \code{data.frame} or \code{matrix} of covariates dataset for testing the association M ~ X and Y ~ M.}

\item{Y.family}{either 'gaussian' (default) or 'binomial', depending on the data type of outcome (\code{Y}). This parameter is passed 
to function \code{lasso.proj} in R package \code{\link{hdi}} for de-biased lasso penalization.}

\item{topN}{an integer specifying the number of top markers from sure independent screening. 
Default = \code{NULL}. If \code{NULL}, \code{topN} will be either \code{ceiling(n/log(n))} if 
\code{Y.family = 'gaussian'}, or \code{ceiling(n/(2*log(n)))} if \code{Y.family = 'binomial'}, 
where \code{n} is the sample size. If the sample size is greater than topN (pre-specified or calculated), all 
mediators will be included in the test (i.e. low-dimensional scenario).}

\item{scale}{logical. Should the function scale the data? Default = \code{TRUE}.}

\item{FDRcut}{FDR cutoff applied to define and select significant mediators. Default = \code{0.05}.}

\item{verbose}{logical. Should the function be verbose? Default = \code{FALSE}.}
}
\value{
A data.frame containing mediation testing results of selected mediators (FDR <\code{FDPcut}). 
\itemize{
    \item{alpha: }{coefficient estimates of exposure (X) --> mediators (M).}
    \item{beta: }{coefficient estimates of mediators (M) --> outcome (Y) (adjusted for exposure).}
    \item{gamma: }{coefficient estimates of exposure (X) --> outcome (Y) (total effect).}
    \item{alpha*beta: }{mediation effect.}
    \item{\% total effect: }{alpha*beta / gamma. Percentage of the mediation effect out of the total effect.}
    \item{p.joint: }{joint p-value of selected significant mediator.}
}

#' @references Perera C, Zhang H, Zheng Y, Hou L, Qu A, Zheng C, Xie K, Liu L. 
HIMA2: high-dimensional mediation analysis and its application in epigenome-wide DNA methylation data. 
BMC Bioinformatics. 2022. DOI: 10.1186/s12859-022-04748-1. PMID: 35879655. PMCID: PMC9310002
}
\description{
\code{dblassoHIMA} is used to estimate and test high-dimensional mediation effects using de-biased lasso penalty.
}
\examples{
\dontrun{
# When Y is continuous and normally distributed
# Example 1 (continuous outcome): 
data(Example1)
head(Example1$PhenoData)

dblassohima.fit <- dblassoHIMA(X = Example1$PhenoData$Treatment, 
                               Y = Example1$PhenoData$Outcome, 
                               M = Example1$Mediator,
                               Z = Example1$PhenoData[, c("Sex", "Age")],
                               Y.family = 'gaussian',
                               scale = FALSE,
                               verbose = TRUE) 
dblassohima.fit

# When Y is binary (should specify Y.family)
# Example 2 (binary outcome): 
data(Example2)
head(Example2$PhenoData)

dblassohima.logistic.fit <- dblassoHIMA(X = Example2$PhenoData$Treatment,
                                        Y = Example2$PhenoData$Disease,
                                        M = Example2$Mediator,
                                        Z = Example2$PhenoData[, c("Sex", "Age")],
                                        Y.family = 'binomial',
                                        scale = FALSE,
                                        verbose = TRUE)
dblassohima.logistic.fit
}

}
