% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eHIMA.R
\name{eHIMA}
\alias{eHIMA}
\title{Efficient high-dimensional mediation analysis}
\usage{
eHIMA(
  X,
  M,
  Y,
  COV = NULL,
  Y.family = c("gaussian"),
  topN = NULL,
  scale = TRUE,
  FDRcut = 0.05,
  verbose = FALSE
)
}
\arguments{
\item{X}{a vector of exposure.}

\item{M}{a \code{data.frame} or \code{matrix} of high-dimensional mediators. Rows represent samples, columns 
represent mediator variables.}

\item{Y}{a vector of continuous outcome. Do not use data.frame or matrix.}

\item{COV}{a matrix of adjusting covariates. Rows represent samples, columns represent variables. Can be \code{NULL}.}

\item{Y.family}{currently \code{eHIMA} only supports 'gaussian', i.e., normally distributed continuous outcome (\code{Y}).}

\item{topN}{an integer specifying the number of top markers from sure independent screening. 
Default = \code{NULL}. If \code{NULL}, \code{topN} will be \code{2*ceiling(n/log(n))}, where \code{n} is the sample size.
If the sample size is greater than topN (pre-specified or calculated), all mediators will be included in the test (i.e. low-dimensional scenario).}

\item{scale}{logical. Should the function scale the data? Default = \code{TRUE}.}

\item{FDRcut}{BH-FDR-corrected p value cutoff applied to define and select significant mediators. Default = \code{0.05}.}

\item{verbose}{logical. Should the function be verbose? Default = \code{FALSE}.}
}
\value{
A data.frame containing mediation testing results of selected mediators (Bonferroni-adjusted p value <\code{Bonfcut}). 
\itemize{
    \item{ID: }{Mediation ID of selected significant mediator.}
    \item{alpha: }{coefficient estimates of exposure (X) --> mediators (M).}
    \item{alpha_se: }{standard error for alpha.}
    \item{beta: }{coefficient estimates of mediators (M) --> outcome (Y) (adjusted for exposure).}
    \item{beta_se: }{standard error for beta.}
    \item{gamma: }{coefficient estimates of exposure (X) --> outcome (Y) (total effect).}
    \item{alpha*beta: }{mediation effect.}
    \item{\% total effect: }{alpha*beta / gamma. Percentage of the mediation effect out of the total effect.}
    \item{p.joint: }{joint raw p-value of selected significant mediator (based on FDR).}
}
}
\description{
\code{eHIMA} is used to estimate and test high-dimensional mediation effects using an efficient algorithm.
}
\examples{
\dontrun{
# Note: In the following example, M1, M2, and M3 are true mediators.
data(himaDat)

# Y is continuous and normally distributed
# Example 1 (continuous outcome): 
head(himaDat$Example1$PhenoData)

eHIMA.fit <- eHIMA(X = himaDat$Example1$PhenoData$Treatment, 
                   Y = himaDat$Example1$PhenoData$Outcome, 
                   M = himaDat$Example1$Mediator,
                   COV = himaDat$Example1$PhenoData[, c("Sex", "Age")],
                   Y.family = 'gaussian',
                   scale = FALSE,
                   verbose = TRUE) 
eHIMA.fit
}

}
\references{
Bai X, Zheng Y, Hou L, Zheng C, Liu L, Zhang H. An Efficient Testing Procedure for High-dimensional Mediators with FDR Control. 
Statistics in Biosciences. 2024. DOI: 10.1007/s12561-024-09447-4.
}
