% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example1.R
\docType{data}
\name{Example1}
\alias{Example1}
\title{Example dataset 1 for HIMA: Continuous outcome}
\format{
An object of class \code{list} of length 2.
}
\usage{
Example1
}
\value{
A list containing two objects: \code{PhenoData} and \code{Mediator}
}
\description{
A \code{list} dataset containing a phenotype dataset and a high-dimension mediator
dataset (n=300 participants, p=300 biomarkers). The variables in the phenotype are as follows:
}
\details{
\itemize{
  \item Treatment: treated (value = 1) or not treated (value = 0)
  \item Outcome: outcome of the treatment- a normally distributed continuous variable
  \item Sex: female (value = 1) or male (value = 0)
  \item Age: Age of the participant
}

The datasets are simulated using parameters generated from a real dataset. The code
used to generate the data can be found in /inst/script folder of the package.
}
\keyword{datasets}
