% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example2.R
\docType{data}
\name{Example2}
\alias{Example2}
\title{Example dataset 2 for HIMA: Binary outcome}
\format{
An object of class \code{list} of length 2.
}
\usage{
Example2
}
\value{
A list containing two objects: \code{PhenoData} and \code{Mediator}
}
\description{
A \code{list} dataset containing a phenotype dataset and a high-dimension mediator
dataset (n=300 participants, p=300 biomarkers). The variables in the phenotype are as follows:
}
\details{
\itemize{
  \item Treatment: treated (value = 1) or not treated (value = 0)
  \item Disease: diseased (value = 1) or healthy (value = 0)
  \item Sex: female (value = 1) or male (value = 0)
  \item Age: Age of the participant
}

The datasets are simulated using parameters generated from a real dataset. The code
used to generate the data can be found in /inst/script folder of the package.
}
\keyword{datasets}
