% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example3.R
\docType{data}
\name{Example3}
\alias{Example3}
\title{Example dataset 3 for HIMA: Survival data outcome}
\format{
An object of class \code{list} of length 2.
}
\usage{
Example3
}
\value{
A list containing two objects: \code{PhenoData} and \code{Mediator}
}
\description{
A \code{list} dataset containing a phenotype dataset and a mediator
dataset (n=200 participants, p=200 biomarkers). The variables in the phenotype are as follows:
}
\details{
\itemize{
  \item Treatment: treated (value = 1) or not treated (value = 0)
  \item Status: Status indicator: dead (value = 1) or alive (value = 0)
  \item Time: time to event
  \item Sex: female (value = 1) or male (value = 0)
  \item Age: Age of the participant
}

The datasets are simulated using parameters generated from a real dataset. The code
used to generate the data can be found in /inst/script folder of the package.
}
\keyword{datasets}
