% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example4.R
\docType{data}
\name{Example4}
\alias{Example4}
\title{Example dataset 4 for HIMA: Compositional mediator (e.g., microbiome)}
\format{
An object of class \code{list} of length 2.
}
\usage{
Example4
}
\value{
A list containing two objects: \code{PhenoData} and \code{Mediator}
}
\description{
A \code{list} dataset containing a phenotype dataset and a compositional mediator
dataset (n=200 participants, p=50 biomarkers). The variables in the phenotype are as follows:
}
\details{
\itemize{
  \item Treatment: treated (value = 1) or not treated (value = 0)
  \item Outcome: outcome of the treatment- a normally distributed continuous variable
  \item Sex: female (value = 1) or male (value = 0)
  \item Age: Age of the participant
}

The datasets are simulated using parameters generated from a real dataset. The code
used to generate the data can be found in /inst/script folder of the package.
}
\keyword{datasets}
