% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HIMA-package.R
\name{HIMA-package}
\alias{HIMA-package}
\title{High-Dimensional Mediation Analysis for 'Omic' Data}
\description{
HIMA is an R package for estimating and testing high-dimensional mediation effects in omic studies.
HIMA can perform high-dimensional mediation analysis on a wide range of omic data types as
potential mediators, including epigenetics, transcriptomics, proteomics, metabolomics, and microbiomics.
HIMA can also handle survival data mediation analysis and perform quantile mediation analysis.

\tabular{ll}{ Package: \tab HIMA\cr Type: \tab Package\cr
Version: \tab 2.3.4\cr Date: \tab 2026-08-30\cr License: \tab GPL-3\cr }
}
\details{
# If package "qvalue" is not found during installation, please first install "qvalue" package
# through Bioconductor: https://www.bioconductor.org/packages/release/bioc/html/qvalue.html
}
\references{
1. Zhang H, Zheng Y, Hou L, Liu L, HIMA: An R Package for High-Dimensional Mediation Analysis. Journal of Data Science. 2025. DOI: 10.6339/25-JDS1192

2. Zhang H, Zheng Y, Zhang Z, Gao T, Joyce B, Yoon G, Zhang W, Schwartz J, Just A, Colicino E, Vokonas P, Zhao L,
Lv J, Baccarelli A, Hou L, Liu L. Estimating and Testing High-dimensional Mediation Effects in Epigenetic Studies.
Bioinformatics. 2016. DOI: 10.1093/bioinformatics/btw351. PMID: 27357171; PMCID: PMC5048064

3. Zhang H, Zheng Y, Hou L, Zheng C, Liu L. Mediation Analysis for Survival Data with High-Dimensional Mediators.
Bioinformatics. 2021. DOI: 10.1093/bioinformatics/btab564. PMID: 34343267; PMCID: PMC8570823

4. Zhang H, Chen J, Feng Y, Wang C, Li H, Liu L. Mediation Effect Selection in High-dimensional and Compositional Microbiome data.
Stat Med. 2021. DOI: 10.1002/sim.8808. PMID: 33205470; PMCID: PMC7855955

5. Zhang H, Chen J, Li Z, Liu L. Testing for Mediation Effect with Application to Human Microbiome Data.
Stat Biosci. 2021. DOI: 10.1007/s12561-019-09253-3. PMID: 34093887; PMCID: PMC8177450

6. Perera C, Zhang H, Zheng Y, Hou L, Qu A, Zheng C, Xie K, Liu L. HIMA2: High-dimensional Mediation Analysis and Its Application in
Epigenome-wide DNA Methylation Data. BMC Bioinformatics. 2022. DOI: 10.1186/s12859-022-04748-1. PMID: 35879655; PMCID: PMC9310002

7. Zhang H, Hong X, Zheng Y, Hou L, Zheng C, Wang X, Liu L. High-Dimensional Quantile Mediation Analysis with Application to a Birth
Cohort Study of Mother–Newborn Pairs. Bioinformatics. 2024. DOI: 10.1093/bioinformatics/btae055. PMID: 38290773; PMCID: PMC10873903

8. Bai X, Zheng Y, Hou L, Zheng C, Liu L, Zhang H. An Efficient Testing Procedure for High-dimensional Mediators with FDR Control.
Statistics in Biosciences. 2024. DOI: 10.1007/s12561-024-09447-4.

9. Liu L, Zhang H, Zheng Y, Gao T, Zheng C, Zhang K, Hou L, Liu L. High-dimensional mediation analysis for longitudinal mediators and 
survival outcomes. Briefings in Bioinformatics. 2025. DOI: 10.1093/bib/bbaf206. PMID: 40350699 PMCID: PMC12066418
}
\author{
Yinan Zheng \email{y-zheng@northwestern.edu},
Haixiang Zhang \email{haixiang.zhang@tju.edu.cn},
Lei liu (Contact) \email{lei.liu@wustl.edu}

Maintainer: Yinan Zheng \email{y-zheng@northwestern.edu}
}
\keyword{package}
