% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hima_rec.R
\name{hima_rec}
\alias{hima_rec}
\title{High-dimensional Mediation Analysis}
\usage{
hima_rec(X, Y, M, COV = NULL, family = c("gaussian", "binomial"),
  penalty = c("MCP", "SCAD", "lasso"), topN = NULL, parallel = TRUE,
  ncore = NULL, verbose = TRUE, maxiter = 5, threshold = 0.05,
  keep.all.first = TRUE, ...)
}
\arguments{
\item{X}{a vector of exposure.}

\item{Y}{a vector of outcome. Can be either continuous or binary (0-1).}

\item{M}{a data frame or matrix of high-dimensional mediators. Rows represent samples, columns 
represent variables.}

\item{COV}{a data frame or matrix of covariates dataset. Covariates specified here will not participate 
penalization. Default = \code{NULL}.}

\item{family}{either 'gaussian' or 'binomial', depending on the data type of outcome (\code{Y}). See 
\code{\link{ncvreg}}}

\item{penalty}{the penalty to be applied to the model. Either 'MCP' (the default), 'SCAD', or 
'lasso'. See \code{\link{ncvreg}}.}

\item{topN}{an integer specifying the number of top markers from sure independent screening. 
Default = \code{NULL}. If \code{NULL}, \code{topN} will be either \code{ceiling(n/log(n))} if 
\code{family = 'gaussian'}, or \code{ceiling(n/(2*log(n)))} if \code{family = 'binomial'}, 
where \code{n} is the sample size.}

\item{parallel}{logical. Enable parallel computing feature? Default = \code{TRUE}.}

\item{ncore}{number of cores to run parallel computation. Valid when \code{parallel == TRUE}. 
By default max number of cores available in the machine will be utilized. If \code{ncore = 1}, 
no parallel computating is allowed.}

\item{verbose}{logical. Should the function be verbose?}

\item{maxiter}{Number of maximum iterations.}

\item{threshold}{Threshold on p-values to be considered as significant.}

\item{keep.all.first}{Do not use threshold on the first round?}

\item{...}{other arguments passed to \code{\link{ncvreg}}.}
}
\value{
The list of markers that are considered significant.
}
\description{
\code{hima_rec} is used to estimate and test high-dimensional mediation effects
by recursively using [hima].
}
\details{
The most significant markers are added as covariables in order to test the 
other markers. This has the benefit to remove markers that are associated to 
a true causal, without being causal for the mediation. It also makes room for
other marker with smaller effects. This procedure stops when there are no 
more significant markers that are detected.
}
