% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/himaFit.R
\name{himaFit}
\alias{himaFit}
\title{High-dimensional Mediation Analysis}
\usage{
himaFit(
  formula,
  data.pheno,
  data.M,
  mediator.type = c("gaussian", "negbin", "compositional"),
  penalty = c("DBlasso", "MCP", "SCAD", "lasso"),
  quantile = FALSE,
  efficient = FALSE,
  scale = TRUE,
  Sigcut = 0.05,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{formula}{an object of class \code{formula}: a symbolic description of the overall effect model, i.e.,
\code{outcome ~ exposure + covariates}, to be fitted. Make sure the "exposure" is the variable of interest, which
must be listed as the first variable in the right hand side of the formula.}

\item{data.pheno}{a data frame containing exposure and covariates that are listed in the right hand side of the \code{formula}.
The variable names must match those listed in \code{formula}. By default \code{himaFit} will scale \code{data.pheno}.}

\item{data.M}{a \code{data.frame} or \code{matrix} of high-dimensional mediators. Rows represent samples, columns
represent variables. By default \code{himaFit} will scale \code{data.M}.}

\item{mediator.type}{data type of high-dimensional mediators (\code{data.M}). Either \code{'gaussian'} (default, for continuous mediators),
\code{'negbin'} (i.e., negative binomial, for RNA-seq data as mediators), or \code{'compositional'} (for microbiome data as mediators).}

\item{penalty}{the penalty to be applied to the model. Either \code{'DBlasso'} (De-biased LASSO, default),
\code{'MCP'}, \code{'SCAD'}, or \code{'lasso'}. Please note, survival HIMA and microbiome HIMA can be only performed with \code{'DBlasso'};
Quantile HIMA cannot be performed with \code{'DBlasso'}. Not applicable for efficient HIMA (when \code{efficient = TRUE}), as it will always apply \code{'MCP'}.}

\item{quantile}{use quantile HIMA (\code{qHIMA}). Only applicable for classic HIMA with continues outcome. Please add parameter \code{tau} to specify the
desired quantile (can be a vector), otherwise the default \code{tau = 0.5} will be used.}

\item{efficient}{use efficient HIMA (\code{eHIMA}). Only applicable for classic HIMA with continuous outcome.
and \code{mediator.type = "gaussian"}). Default = \code{FALSE}.}

\item{scale}{logical. Should the function scale the data (exposure, mediators, and covariates)? Default = \code{TRUE}. Note: for simulation study, scale
can be turned off to avoid estimate compression.}

\item{Sigcut}{cutoff applied to select significant mediators. Default = \code{0.05}.}

\item{verbose}{logical. Should the function be verbose and show the progression? Default = \code{FALSE}.}

\item{...}{reserved passing parameter.}
}
\value{
A data.frame containing mediation testing results of selected mediators.
\describe{
    \item{ID: }{Mediator ID/name.}
    \item{alpha: }{Coefficient estimates of exposure (X) --> mediators (M) (adjusted for covariates).}
    \item{beta: }{Coefficient estimates of mediators (M) --> outcome (Y) (adjusted for covariates and exposure).}
    \item{alpha*beta: }{Mediation (indirect) effect.}
    \item{Relative Importance: }{Relative importance of the mediator. It is the proportion of the mediation effect for each mediator
    out of the sum of the mediation effect (absolute value) across all significant mediators selected.}
    \item{p-value: }{Joint raw p-value of significant mediators selected based on corresponding approach.}
    \item{tau: }{Quantile level of the outcome (applicable only to the quantile mediation model).}
}
}
\description{
\code{himaFit} is a wrapper function designed to perform various HIMA methods for estimating and testing high-dimensional mediation effects.
\code{himaFit} can automatically select the appropriate HIMA method based on the outcome and mediator data type.
}
\examples{
\dontrun{
# Note: In the following examples, M1, M2, and M3 are true mediators.
data(himaDat)

# Example 1 (continuous outcome - linear HIMA):
head(himaDat$Example1$PhenoData)

e1 <- himaFit(Outcome ~ Treatment + Sex + Age,
  data.pheno = himaDat$Example1$PhenoData,
  data.M = himaDat$Example1$Mediator,
  mediator.type = "gaussian",
  penalty = "MCP", # Can be "DBlasso" for dblassoHIMA
  scale = FALSE
) # Disabled only for simulation data
e1
attributes(e1)$variable.labels

# Efficient HIMA (only applicable to mediators and outcomes that are
# both continuous and normally distributed.)
e1e <- himaFit(Outcome ~ Treatment + Sex + Age,
  data.pheno = himaDat$Example1$PhenoData,
  data.M = himaDat$Example1$Mediator,
  mediator.type = "gaussian",
  efficient = TRUE,
  scale = FALSE
) # Disabled only for simulation data
e1e
attributes(e1e)$variable.labels

# Example 2 (binary outcome - logistic HIMA):
head(himaDat$Example2$PhenoData)

e2 <- himaFit(Disease ~ Treatment + Sex + Age,
  data.pheno = himaDat$Example2$PhenoData,
  data.M = himaDat$Example2$Mediator,
  mediator.type = "gaussian",
  penalty = "MCP",
  scale = FALSE
) # Disabled only for simulation data
e2
attributes(e2)$variable.labels

# Example 3 (time-to-event outcome - survival HIMA):
head(himaDat$Example3$PhenoData)

e3 <- himaFit(Surv(Status, Time) ~ Treatment + Sex + Age,
  data.pheno = himaDat$Example3$PhenoData,
  data.M = himaDat$Example3$Mediator,
  mediator.type = "gaussian",
  penalty = "DBlasso",
  scale = FALSE
) # Disabled only for simulation data
e3
attributes(e3)$variable.labels

# Example 4 (compositional data as mediator, e.g., microbiome):
head(himaDat$Example4$PhenoData)

e4 <- himaFit(Outcome ~ Treatment + Sex + Age,
  data.pheno = himaDat$Example4$PhenoData,
  data.M = himaDat$Example4$Mediator,
  mediator.type = "compositional",
  penalty = "DBlasso"
) # Scaling is always enabled internally for microHIMA
e4
attributes(e4)$variable.labels

#' # Example 5 (quantile mediation anlaysis - quantile HIMA):
head(himaDat$Example5$PhenoData)

# Note that the function will prompt input for quantile level.
e5 <- himaFit(Outcome ~ Treatment + Sex + Age,
  data.pheno = himaDat$Example5$PhenoData,
  data.M = himaDat$Example5$Mediator,
  mediator.type = "gaussian",
  quantile = TRUE,
  penalty = "MCP", # Quantile HIMA does not support DBlasso
  scale = FALSE, # Disabled only for simulation data
  tau = c(0.3, 0.5, 0.7)
) # Specify multiple quantile level
e5
attributes(e5)$variable.labels
}

}
\references{
1. Zhang H, Zheng Y, Zhang Z, Gao T, Joyce B, Yoon G, Zhang W, Schwartz J, Just A, Colicino E, Vokonas P, Zhao L,
Lv J, Baccarelli A, Hou L, Liu L. Estimating and Testing High-dimensional Mediation Effects in Epigenetic Studies.
Bioinformatics. 2016. DOI: 10.1093/bioinformatics/btw351. PMID: 27357171; PMCID: PMC5048064

2. Zhang H, Zheng Y, Hou L, Zheng C, Liu L. Mediation Analysis for Survival Data with High-Dimensional Mediators.
Bioinformatics. 2021. DOI: 10.1093/bioinformatics/btab564. PMID: 34343267; PMCID: PMC8570823

3. Zhang H, Chen J, Feng Y, Wang C, Li H, Liu L. Mediation Effect Selection in High-dimensional and Compositional Microbiome data.
Stat Med. 2021. DOI: 10.1002/sim.8808. PMID: 33205470; PMCID: PMC7855955

4. Zhang H, Chen J, Li Z, Liu L. Testing for Mediation Effect with Application to Human Microbiome Data.
Stat Biosci. 2021. DOI: 10.1007/s12561-019-09253-3. PMID: 34093887; PMCID: PMC8177450

5. Perera C, Zhang H, Zheng Y, Hou L, Qu A, Zheng C, Xie K, Liu L. HIMA2: High-dimensional Mediation Analysis and Its Application in
Epigenome-wide DNA Methylation Data. BMC Bioinformatics. 2022. DOI: 10.1186/s12859-022-04748-1. PMID: 35879655; PMCID: PMC9310002

6. Zhang H, Hong X, Zheng Y, Hou L, Zheng C, Wang X, Liu L. High-Dimensional Quantile Mediation Analysis with Application to a Birth
Cohort Study of Mother–Newborn Pairs. Bioinformatics. 2024. DOI: 10.1093/bioinformatics/btae055. PMID: 38290773; PMCID: PMC10873903

7. Bai X, Zheng Y, Hou L, Zheng C, Liu L, Zhang H. An Efficient Testing Procedure for High-dimensional Mediators with FDR Control.
Statistics in Biosciences. 2024. DOI: 10.1007/s12561-024-09447-4.
}
