% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/himaDat.R
\docType{data}
\name{MicrobiomeData}
\alias{MicrobiomeData}
\title{Compositional Mediator Dataset for HIMA Demo}
\format{
A list with the following components:
\describe{
  \item{PhenoData}{A data frame containing:
    \describe{
      \item{Treatment}{treated (value = 1) or not treated (value = 0).}
      \item{Outcome}{a normally distributed continuous outcome variable.}
      \item{Sex}{female (value = 1) or male (value = 0).}
      \item{Age}{age of the participant.}
    }
  }
  \item{Mediator}{A matrix of high-dimensional compositional mediators (rows: samples, columns: variables).}
}
}
\usage{
MicrobiomeData
}
\description{
A dataset containing phenotype data and high-dimensional compositional mediators (e.g., microbiome).
The dataset was simulated using parameters generated from real data.
}
\examples{
data(MicrobiomeData)
head(MicrobiomeData$PhenoData)

}
\keyword{datasets}
