% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/himaDat.R
\docType{data}
\name{QuantileData}
\alias{QuantileData}
\title{Quantile Mediation Dataset for HIMA Demo}
\format{
A list with the following components:
\describe{
  \item{PhenoData}{A data frame containing:
    \describe{
      \item{Treatment}{treated (value = 1) or not treated (value = 0).}
      \item{Outcome}{an abnormally distributed continuous outcome variable.}
      \item{Sex}{female (value = 1) or male (value = 0).}
      \item{Age}{age of the participant.}
    }
  }
  \item{Mediator}{A matrix of high-dimensional mediators (rows: samples, columns: variables).}
}
}
\usage{
QuantileData
}
\description{
A dataset containing phenotype data and high-dimensional mediators for quantile mediation analysis.
The dataset was simulated using parameters generated from real data.
}
\examples{
data(QuantileData)
head(QuantileData$PhenoData)

}
\keyword{datasets}
