% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simHIMA.R
\name{simHIMA}
\alias{simHIMA}
\title{Simulation Data Generator for High-dimensional Mediation Analyais}
\usage{
simHIMA(n, p, alpha, beta, seed)
}
\arguments{
\item{n}{an integer specifying sample size.}

\item{p}{an integer specifying the dimension of mediators.}

\item{alpha}{a numeric vector specifying the regression coefficients alpha (exposure --> mediators).}

\item{beta}{a numeric vector specifying the regression coefficients beta (mediators --> outcome).}

\item{seed}{an integer specifying a seed for random number generation.}
}
\description{
\code{simHIMA} is used to generate simulation data for high-dimensional mediation analyais.
}
\examples{
n <- 300  # sample size
p <- 10000 # the dimension of covariates

alpha <- rep(0, p) # the regression coefficients alpha (exposure --> mediators)
beta <- rep(0, p) # the regression coefficients beta (mediators --> outcome)
alpha[1:4] <- c(0.45,0.55,0.5,0.55)
beta[1:4] <- c(0.40,0.55,0.5,0.40) # the first four makers are true mediators.

alpha[7:8] <- 0.55
beta[5:6] <- 0.50

simdat = simHIMA(n, p, alpha, beta, seed=2016) # Generate simulation data

}
\seealso{
see \code{\link{hima}} to run HIMA.
}

