% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simHMA.R
\name{simHMA}
\alias{simHMA}
\title{Simulation Data Generator for High-dimensional Mediation Analyais}
\usage{
simHMA(n, p, a, beta)
}
\arguments{
\item{n}{an integer specifying sample size.}

\item{p}{an integer specifying the dimension of covariates.}

\item{beta}{a numeric vector specifying the regression coefficients beta}

\item{alpha}{a numeric vector specifying the regression coefficients alpha}
}
\description{
\code{simHMA} is used to simulate data for High-dimensional Mediation Analyais
}
\examples{
n <- 100  # sample size
p <- 5000 # the dimension of covariates

alpha <- rep(0, p) # the regression coefficients alpha
beta <- rep(0, p) # the regression coefficients beta

alpha[1:3]  <- c(0.5, 0.5, 0.3)
beta[1:3] <- c(0.5, 1.2, 0.3)

simdat = simHMA(n, p, alpha, beta)
}
\seealso{
see \code{\link{hma}} to run HMA algorithm.
}

