% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sis.R
\name{sis}
\alias{sis}
\title{Sure Independent Screening}
\usage{
sis(Y, x.vect, COV = NULL, parallel = TRUE, ncore = 2)
}
\arguments{
\item{Y}{a data frame or matrix of genomic dataset. Rows represent samples, columns represent variables.}

\item{x.vect}{a vector of independent variable.}

\item{COV}{a data frame or matrix of covariates dataset. Default = \code{NULL}.}

\item{parallel}{logical. Enable parallel computing feature? Default = \code{TRUE}.}

\item{ncore}{number of cores to run parallel. Effective when paralle = \code{TRUE}. Default = 2.}
}
\value{
a vector of estimates of mediators in their original order.
}
\description{
\code{sis} is used to conduct sure independent screening using linear regression across high-dimensional variables as outcomes.
}
\examples{

n <- 100  # sample size
p <- 5000 # the dimension of covariates

alpha <- rep(0, p) # the regression coefficients alpha (exposure --> mediators)
beta <- rep(0, p) # the regression coefficients beta (mediators --> outcome)
alpha[1:3]  <- c(0.5, 0.5, 0.3) 
beta[1:3] <- c(0.5, 1.2, 0.3) # the first three makers are true mediators.
intercept <- runif(p,0,2)

set.seed(2016)
simdat = simHIMA(n, p, alpha, beta, intercept)
}
\seealso{
see \code{\link{hima}} using the results from \code{\link{sis}} as input to run PGS.
}

