\name{HIMA-package}
\alias{HIMA-package}
\alias{HIMA}
\docType{package}
\title{
\packageTitle{HIMA}
}
\description{
\packageDescription{HIMA}
}
\details{
\code{\link{hima}} is used to estimate and test high-dimensional mediation effects.

\code{\link{simHIMA}} is used to simulate data for high-dimensional mediation analyais.
}
\author{
\packageAuthor{HIMA}

Maintainer: \packageMaintainer{HIMA}
}
\keyword{package}