% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hima.R
\name{hima}
\alias{hima}
\title{High-dimensional Mediation Analysis}
\usage{
hima(X, Y, M, COV = NULL, family = c("gaussian", "binomial"),
  penalty = c("MCP", "SCAD", "lasso"), topN = NULL, parallel = TRUE,
  ncore = NULL, verbose = TRUE, ...)
}
\arguments{
\item{X}{a vector of exposure.}

\item{Y}{a vector of outcome. Can be either continuous or binary (0-1).}

\item{M}{a data frame or matrix of high-dimensional mediators. Rows represent samples, columns 
represent variables.}

\item{COV}{a data frame or matrix of covariates dataset. Default = \code{NULL}.}

\item{family}{either 'gaussian' or 'binomial', depending on the data type of outcome (\code{Y}). See 
\code{\link{ncvreg}}}

\item{penalty}{the penalty to be applied to the model. Either 'MCP' (the default), 'SCAD', or 
'lasso'. See \code{\link{ncvreg}}.}

\item{topN}{an integer specifying the number of top markers from sure independent screening. 
Default = \code{NULL}. If \code{NULL}, \code{ceiling(n/log10(n))} will be used as \code{topN}, 
where \code{n} is the sample size.}

\item{parallel}{logical. Enable parallel computing feature? Default = \code{TRUE}.}

\item{ncore}{number of cores to run parallel computation. Valid when \code{parallel == TRUE}. 
By default max number of cores available in the machine will be utilized. If \code{ncore = 1}, 
no parallel computating is allowed.}

\item{verbose}{logical. Should the function be verbose?}

\item{...}{other arguments passed to \code{\link{ncvreg}}.}
}
\value{
A data.frame containing mediation testing results of selected mediators.
}
\description{
\code{hima} is used to estimate and test high-dimensional mediation effects.
}
\examples{
n <- 300  # sample size
p <- 10000 # the dimension of covariates

# the regression coefficients alpha (exposure --> mediators)
alpha <- rep(0, p) 

# the regression coefficients beta (mediators --> outcome)
beta1 <- rep(0, p) # for continuous outcome
beta2 <- rep(0, p) # for binary outcome
# the first four markers are true mediators
alpha[1:4] <- c(0.45,0.5,0.55,0.6)
beta1[1:4] <- c(0.5,0.45,0.4,0.35)
beta2[1:4] <- c(1.5,1.45,1.4,1.35)

# these are also not true mediators
alpha[7:8] <- 0.5
beta1[5:6] <- 0.4
beta2[5:6] <- 1.4

# Generate simulation data
simdat_cont = simHIMA(n, p, alpha, beta1, seed=2016) 
simdat_bin = simHIMA(n, p, alpha, beta2, binaryOutcome = TRUE, seed=2016) 
# Run HIMA with MCP penalty by default
hima.fit <- hima(simdat_cont$X, simdat_cont$Y, simdat_cont$M, 
parallel = FALSE) 
head(hima.fit)

hima.logistic.fit <- hima(simdat_bin$X, simdat_bin$Y, simdat_bin$M, 
family = "binomial", , parallel = FALSE) 
head(hima.logistic.fit)

}

