% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simHIMA.R
\name{simHIMA}
\alias{simHIMA}
\title{Simulation Data Generator for High-dimensional Mediation Analyais}
\usage{
simHIMA(n, p, alpha, beta, intercept)
}
\arguments{
\item{n}{an integer specifying sample size.}

\item{p}{an integer specifying the dimension of covariates.}

\item{alpha}{a numeric vector specifying the regression coefficients alpha.}

\item{beta}{a numeric vector specifying the regression coefficients beta.}

\item{c}{a numeric vector specifying intercept of each mediator.}
}
\description{
\code{simHIMA} is used to simulate data for High-dimensional Mediation Analyais
}
\examples{
n <- 100  # sample size
p <- 5000 # the dimension of covariates

alpha <- rep(0, p) # the regression coefficients alpha
beta <- rep(0, p) # the regression coefficients beta

alpha[1:3]  <- c(0.5, 0.5, 0.3)
beta[1:3] <- c(0.5, 1.2, 0.3)
intercept <- runif(p,0,2)

simdat = simHMA(n, p, alpha, beta, c)
}
\seealso{
see \code{\link{hima}} to run HMA algorithm.
}

