% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simHIMA.R
\name{simHIMA}
\alias{simHIMA}
\title{Simulation Data Generator for High-dimensional Mediation Analyais}
\usage{
simHIMA(n, p, alpha, beta, seed)
}
\arguments{
\item{n}{an integer specifying sample size.}

\item{p}{an integer specifying the dimension of mediators.}

\item{alpha}{a numeric vector specifying the regression coefficients alpha (exposure --> mediators).}

\item{beta}{a numeric vector specifying the regression coefficients beta (mediators --> outcome).}

\item{seed}{an integer specifying a seed for random number generation.}
}
\description{
\code{simHIMA} is used to generate simulation data for high-dimensional mediation analyais.
}
\examples{
n <- 100  # sample size
p <- 5000 # the dimension of covariates

alpha <- rep(0, p) # the regression coefficients alpha (exposure --> mediators)
beta <- rep(0, p) # the regression coefficients beta (mediators --> outcome)
alpha[1:3] <- c(0.5, 0.5, 0.3) 
beta[1:3] <- c(0.5, 1.2, 0.3) # the first three makers are true mediators.

simdat = simHIMA(n, p, alpha, beta, seed=2016) # Generate simulation data

}
\seealso{
see \code{\link{hima}} to run HIMA.
}

