% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survHIMA.R
\name{survHIMA}
\alias{survHIMA}
\title{High-dimensional mediation analysis for survival data}
\usage{
survHIMA(X, Z, M, y, status, FDRcut = 0.05)
}
\arguments{
\item{X}{a vector of exposure.}

\item{Z}{a matrix of adjusting covariates. Rows represent samples, columns represent variables.}

\item{M}{a \code{data.frame} or \code{matrix} of high-dimensional mediators. Rows represent samples, columns 
represent variables.}

\item{y}{a vector of observed failure times.}

\item{status}{a vector of censoring indicator (\code{status = 1}: uncensored; \code{status = 0}: censored)}

\item{FDRcut}{FDR cutoff applied to define and select significant mediators. Default = \code{0.05}.}
}
\value{
A data.frame containing mediation testing results of selected mediators (FDR <\code{0.05}). 
\itemize{
    \item{ID: }{index of selected significant mediator.}
    \item{alpha: }{coefficient estimates of exposure (X) --> mediators (M).}
    \item{alpha_se: }{standard error for alpha.}
    \item{beta: }{coefficient estimates of mediators (M) --> outcome (Y) (adjusted for exposure).}
    \item{beta_se: }{standard error for beta}
    \item{p.joint: }{joint p-value of selected significant mediator.}
}
}
\description{
\code{survHIMA} is used to estimate and test high-dimensional mediation effects for survival data.
}
\examples{
## Generate simulated survival data
n <- 300  # sample size
c_unif <- 150 # CR= 0.20

##
p <- 100 # the dimension of mediators
q <- 2   # the dimension of covariates
##

sigma_e <- matrix(0.25, p, p)
diag(sigma_e) <- 1
sigma_e[1, 3] <- 0.8
sigma_e[3, 1] <- sigma_e[1, 3]
sigma_e[2, 4] <- 0.8
sigma_e[4, 2] <- 0.8

##
beta <- matrix(0, 1, p)
beta[1]  <- 0.6
beta[3]  <- 0.8
beta[5]  <- 0.55

##
alpha <- matrix(0, 1, p)
alpha[1]  <- 0.6
alpha[3]  <- 0.8
alpha[5]  <- 0.8

##
gamma <- matrix(0.5, 1, q)
eta <- matrix(0.3, p, q)
r <- matrix(0.5, 1, 1)

##
X <- matrix(rnorm(n, mean = 0, sd = 2), n, 1) # expoure
Z <- matrix(rnorm(n * q, mean = 0, sd = 2), n, q) # covariates
mu <- matrix(0, p, 1)
e <- MASS::mvrnorm(n, mu, sigma_e)  # the error terms

M <- matrix(0, n, p)
M <- X\%*\%(alpha) + Z\%*\%t(eta) + e
MZ <- cbind(M, Z, X)

beta_gamma <- cbind(beta,gamma,r)

## generate the filure time T 
u <- runif(n, 0, 1)
T <- matrix(0, n, 1) 
for (i in 1:n){
  T[i] <- -log(1 - u[i])*exp(-sum(beta_gamma*MZ[i,]))
}

## generate censoring time 0.45 censoring rate
C <- runif(n, min = 0, max = c_unif)  

status <- as.integer(T < C)

## the observed failure time
y <- apply(cbind(C,T), 1, min) 

fit <- survHIMA(X, Z, M, y, status)

## ID = 1, 3, 5, are selected (FDR <0.05)
fit

}
